+++ packages/davecode-events/src/index.ts
@@ -1,24 +1,24 @@
-export type BaseEventMap = Record<string, any> | {};
-export type EventHandler<T> = undefined | never extends T
-  ? (event: never) => void
-  : (event: T) => void;
+import type { Dict, EmptyObject } from '@davecode/types';
 
-type UndefinedKeys<T> = {} extends T
-  ? string
-  : { [K in keyof T]: undefined | never extends T[K] ? K : never }[keyof T];
-type DefinedKeys<T> = {} extends T
-  ? string
-  : { [K in keyof T]: undefined | never extends T[K] ? never : K }[keyof T];
-type AllKeys<T> = {} extends T ? string : keyof T;
+export type BaseEventMap = Record<string, any> | Dict<unknown>;
+export type EventHandler<T> = undefined extends T ? (event: never) => void : (event: T) => void;
 
-// @ts-ignore
+type UndefinedKeys<T> = EmptyObject extends T
+  ? string
+  : { [K in keyof T]: undefined extends T[K] ? K : never }[keyof T];
+type DefinedKeys<T> = Dict<unknown> extends T
+  ? string
+  : { [K in keyof T]: undefined extends T[K] ? never : K }[keyof T];
+type AllKeys<T> = Dict<unknown> extends T ? string : keyof T;
+
+// @ts-expect-error
 type Key<T, K extends PropertyKey> = T[K];
 
 /**
  * A generic typed event-emitter class. Only supports singular argument events. Pass an interface
  * mapping event names to their event value.
  */
-export class Emitter<T extends BaseEventMap = {}> {
+export class Emitter<T extends BaseEventMap = Dict<unknown>> {
   #listeners: Partial<Record<AllKeys<T>, Set<EventHandler<any>>>> = {};
 
   /** Adds an event listener. Returns a function that can be called to remove the listener. */
@@ -39,7 +39,7 @@ export class Emitter<T extends BaseEventMap = {}> {
   emit<K extends UndefinedKeys<T>>(event: K): void;
   emit<K extends DefinedKeys<T>>(event: K, data: Key<T, K>): void;
   emit<K extends keyof T>(event: K, data?: T[K]): void {
-    this.#listeners[event]?.forEach(listener => (listener as any)(data));
+    this.#listeners[event]?.forEach((listener: any) => listener(data));
   }
 
   /** Removes all listeners for an event or all if no event is provided. */
@@ -60,6 +60,6 @@ export class Emitter<T extends BaseEventMap = {}> {
 
   /** Removes an event listener by given event name and listener function. Identical to `.off` */
   removeEventListener<K extends AllKeys<T>>(event: K, listener: EventHandler<Key<T, K>>) {
-    return this.off(event, listener);
+    this.off(event, listener);
   }
 }