+++ packages/davecode-events/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Patch Changes
 
+- fix types a little bit
+
+## 1.0.1
+
+### Patch Changes
+
 - add more utils
 - Updated dependencies [176570b]
   - @davecode/types@1.0.2